# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Good Food Tracker API",
    "description": "The project aims to allow the users to take pictures and/or leave notes, ratings, comments about restaurants they visit\r\n\tin order to be able to reference it later when they try to pick were they want to go eat out or order from.",
    "contact": {
      "name": "Igor Ilić",
      "email": "github@igorilic.net"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8098",
      "description": "Development API base url"
    },
    {
      "url": "https://gft.igorilic.net/api",
      "description": "Production API base url"
    }
  ],
  "paths": {
...
    "/city": {
      "get": {
        "tags": [
          "City"
        ],
        "summary": "Fetch a list of cities",
        "description": "Endpoint used for getting a list of cities",
        "operationId": "e8d501fe8bfdc517ca4673c0d2a0e768",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Value used to filter the cities",
            "required": false,
            "schema": {
              "type": "string",
              "additionalProperties": false
            }
          },
          {
            "name": "start",
            "in": "query",
            "description": "Pagination start offset",
            "required": false,
            "schema": {
              "type": "integer",
              "additionalProperties": false
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Pagination end offset",
            "required": false,
            "schema": {
              "type": "integer",
              "additionalProperties": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CityEntity"
                      }
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "City"
        ],
        "summary": "Create a city",
        "description": "Endpoint used for creating a city",
        "operationId": "adacc285c00288c4c0bf045b5839f6c5",
        "requestBody": {
          "description": "Required parameters",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "country_id": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CityEntity"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Missing required arguments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/city/{cityID}": {
      "get": {
        "tags": [
          "City"
        ],
        "summary": "Fetch a single city",
        "description": "Endpoint used for get information about a single city",
        "operationId": "c3217ec998b12b7ced2316cdce2a1d6f",
        "parameters": [
          {
            "name": "cityID",
            "in": "path",
            "description": "ID of a city we are looking for",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/uuid_parameter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CityEntity"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Missing required arguments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "404": {
            "description": "City not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "City"
        ],
        "summary": "Delete city",
        "description": "Endpoint used for deleting a city",
        "operationId": "3728f52c10130ab55512ed04c7db7ee7",
        "parameters": [
          {
            "name": "cityID",
            "in": "path",
            "description": "ID of a city being updated",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/uuid_parameter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CityEntity"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Missing required arguments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "City"
        ],
        "summary": "Update city",
        "description": "Endpoint used for updating city information",
        "operationId": "4fd8a9400071d1b95c78581570f67d57",
        "parameters": [
          {
            "name": "cityID",
            "in": "path",
            "description": "ID of a city being updated",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/uuid_parameter"
            }
          }
        ],
        "requestBody": {
          "description": "Required parameters",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CityEntity"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Missing required arguments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CityEntity": {
        "description": "CityEntity Class",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "country_id": {
            "type": "string"
          },
          "country_name": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "error_response": {
        "properties": {
          "class": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "field": {
            "type": "string"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
...
      "uuid_parameter": {
        "description": "Swagger doc block",
        "type": "string",
        "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[1-5]{1}[0-9A-Fa-f]{3}-[ABab89]{1}[0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}$",
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}